// packages/client/src/store/actions/types.ts
export type Action = { type: ActionType; payload: any };
export enum ActionType {
  SET_ALERT = "SET_ALERT",
  ANIMATE_ALERT = "ANIMATE_ALERT",
  REMOVE_ALERT = "REMOVE_ALERT",
  REGISTER_SUCCESS = "REGISTER_SUCCESS",
  REGISTER_FAIL = "REGISTER_FAIL",
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  LOGIN_FAIL = "LOGIN_FAIL",
  USER_LOADED = "USER_LOADED",
  AUTH_ERROR = "AUTH_ERROR",
  LOGOUT = "LOGOUT",
  SET_NEW_CHAT_ROOM_LOADING = "SET_NEW_CHAT_ROOM_LOADING",
  UPDATE_CURRENT_CHAT_ROOM = "UPDATE_CURRENT_CHAT_ROOM",
  NEW_CHAT_MESSAGE = "NEW_CHAT_MESSAGE",
  VIEW_GAMES_LIST_CLICKED = "VIEW_GAMES_LIST_CLICKED",
  UPDATE_GAMES_LIST = "UPDATE_GAMES_LIST",
  CLOSE_GAME_LIST = "CLOSE_GAME_LIST",
  OPEN_PRE_GAME_SCREEN = "OPEN_PRE_GAME_SCREEN",
  CLOSE_PRE_GAME_SCREEN = "CLOSE_PRE_GAME_SCREEN",
  SET_CURRENT_GAME = "SET_CURRENT_GAME",
  UPDATE_PLAYERS = "UPDATE_PLAYERS",
  UPDATE_PLAYERS_READY = "UPDATE_PLAYERS_READY",
  UPDATE_GAME_COUNTDOWN = "UPDATE_GAME_COUNTDOWN",
  UPDATE_GAME_STATUS = "UPDATE_GAME_STATUS",
  UPDATE_PLAYER_ROLE = "UPDATE_PLAYER_ROLE",
  SET_GAME_WINNER = "SET_GAME_WINNER",
  CLEAR_GAME_UI = "CLEAR_GAME_UI",
  SET_SCORE_SCREEN_DATA = "SET_SCORE_SCREEN_DATA",
  CLOSE_SCORE_SCREEN = "CLOSE_SCORE_SCREEN",
  GET_LADDER_PAGE = "GET_LADDER_PAGE",
  CHANGE_LADDER_PAGE_VIEWING = "CHANGE_LADDER_PAGE_VIEWING",
  GET_BATTLE_ROOM_USER_RECORD = "GET_BATTLE_ROOM_USER_RECORD",
  SET_MATCHMAKING_DATA = "SET_MATCHMAKING_DATA",
  SET_MATCHMAKING_WINDOW_VISIBLE = "SET_MATCHMAKING_WINDOW_VISIBLE",
}